sesion_05/requirements.txt
Django>=3.1.2,<3.2.0
djangorestframework>=3.12.1,<3.13.0
markdown>=3.3,<3.4
django-filter>=2.4.0,<2.5.0
